feat: 소수점 처리 기능 구현

+++ src/main/java/calculator/Application.java
@@ -1,15 +1,15 @@
 package calculator;
 
-import java.math.BigInteger;
+import java.math.BigDecimal;
 
 public class Application {
     public static void main(String[] args) {
-        
+
         ConsoleIO consoleIO = new ConsoleIO();
         String input = consoleIO.getInput();
 
         StringCalculator stringCalculator = new StringCalculator();
-        BigInteger result = stringCalculator.calculateString(input);
+        BigDecimal result = stringCalculator.calculateString(input);
 
         consoleIO.displayOutput(result);
     }
